feat(core): add convert_polynomials_list_to_fourier
